+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -5,7 +5,8 @@ Release History
 
 2.0.36
 ++++++
-* Minor fixes
+* `vm/vmss extension set/delete`: Added `--no-wait` support.
+* Added `vm extension wait`.
 
 2.0.35
 ++++++